requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
aiofiles==24.1.0
pillow==11.1.0
numpy==1.26.4
h5py==3.12.1
tensorflow==2.16.1
